requirements.txt
streamlit
pdfplumber
python-docx
flashtext
pytesseract
Pillow
pandas
spacy
